requirements.txt
requests
python-telegram-bot
schedule
matplotlib
pandas
python-dotenv
